unify oslogin packaging

+++ google_compute_engine_oslogin/packaging/setup_deb.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# Copyright 2017 Google Inc. All Rights Reserved.
+# Copyright 2018 Google Inc. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,26 +13,23 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-############################## WARNING ##############################
-# This script is for testing purposes only. It is not intended
-# for creating production Debian packages.
-#####################################################################
-
-# Run from the top of the source directory.
 NAME="google-compute-engine-oslogin"
 VERSION="1.4.3"
 
 working_dir=${PWD}
+if [[ $(basename "$working_dir") != $NAME ]]; then
+  echo "Packaging scripts must be run from top of package dir."
+  exit 1
+fi
 
 # Build dependencies.
 sudo apt-get -y install make g++ libcurl4-openssl-dev libjson-c-dev libpam-dev
 
-# .deb creation tools.
+# DEB creation tools.
 sudo apt-get -y install debhelper devscripts build-essential
 
 rm -rf /tmp/debpackage
-mkdir -p /tmp/debpackage/${NAME}-${VERSION}
-
+mkdir /tmp/debpackage
 tar czvf /tmp/debpackage/${NAME}_${VERSION}.orig.tar.gz  --exclude .git --exclude packaging --transform "s/^\./${NAME}-${VERSION}/" .
 
 pushd /tmp/debpackage
@@ -40,8 +37,7 @@ tar xzvf ${NAME}_${VERSION}.orig.tar.gz
 
 cd ${NAME}-${VERSION}
 
-deb_version=$(cut -d "." -f 1 /etc/debian_version)
-cp -r ${working_dir}/packaging/debian${deb_version} debian
+cp -r ${working_dir}/packaging/debian ./
 
 debuild -us -uc
